Добавил логирование

+++ 7.AspnetMvc/Lesson-7/FileCommander.Services/FileService.cs
@@ -26,31 +26,42 @@ public class FileService : IFileService
     {
         var report = _reportResolver.GetReportByFileName(sourceFileName);
         var options = new CreateReportOptions() {  OpenAfterCreate = true};
+
+        _logger.LogInformation("Создание отчета <{0}> \"{1}\" для объекта \"{2}\"", report.GetType().Name , reportFileName, sourceFileName);
         report.MakeReport(reportFileName, sourceFileName, options);
     }
 
     public async Task CopyFileAsync(string source, string destination, IProgress<int>? progress = default, CancellationToken cancel = default)
     {
         cancel.ThrowIfCancellationRequested();
-        using FileStream reader = File.OpenRead(source);
-        int bufferSize = reader.Length < DefaultBufferSize ? Convert.ToInt32(reader.Length) : DefaultBufferSize;
-        var buffer = new byte[bufferSize];
-        using FileStream writer = File.OpenWrite(destination);
+      
+        if (!File.Exists(source))
+            throw new FileNotFoundException($"Файл {source} не найден");
+
+        _logger.LogInformation("Копирование файла {0} в папку {1}...", source, destination);
         int read = 0;
         long totalRead = 0;
         int calculatedProgress = 0;
         int currentProgress = 0;
+
+        using FileStream reader = File.OpenRead(source);
+        int bufferSize = reader.Length < DefaultBufferSize ? Convert.ToInt32(reader.Length) : DefaultBufferSize;
+        var buffer = new byte[bufferSize];
+
+        using FileStream writer = File.OpenWrite(destination);
         while ((read = await reader.ReadAsync(buffer, 0, bufferSize, cancel).ConfigureAwait(false)) > 0)
         {
             if (cancel.IsCancellationRequested)
             {
                 writer.Close();
                 File.Delete(destination);
-                _logger.LogInformation("Отмена операции копирования файла {0} в {1}", source, destination);
+                _logger.LogInformation("Отмена операции копирования файла {0} в папку {1}", source, destination);
                 throw new OperationCanceledException();
             }
+
             await writer.WriteAsync(buffer, 0, read, cancel).ConfigureAwait(false);
             totalRead += read;
+
             calculatedProgress = (int)((double)totalRead / reader.Length * 100);
             if (currentProgress != calculatedProgress)
             {
@@ -58,5 +69,6 @@ public class FileService : IFileService
                 progress?.Report(currentProgress);
             }
         }
+        _logger.LogInformation("Файл успешно {0} успешно скопирован в папку {1}", source, destination);
     }
 }
